Add explanatory comment
I was worried that this was buggy but I think it’s okay.
If there happens to be no gap between the last build
loaded via the builds API and the last build on the
default branch, there’ll just be a slight overlap on
the new builds API call.