Fix gap in build history (#865)
This is a workaround for some weirdness that happens when
a repository’s default branch hasn’t had a build in a while.
If it’s been more than the number of builds fetched for the
build history tab (20), there’s a gap in the displayed builds.
When you clicked “Show more”, you didn’t get the builds
between the 20th recent build and the default branch’s
last build.

This adds a hack to address that. It’s possible that
the default branch’s build shouldn’t even show at
all if it’s not within the most recent builds, but if so
we can come back to this.