Fix sign in with ember server and redirect flow (#900)
Recently GitHub introduced Content Security Policy header that disallows
putting GitHub in an iframe. That's why our iframe sign in flow fails on
browsers that support CSP headers. On top of that the redirect flow that
we used as a fallback didn't work properly when using ember-cli's built
in server, because it didn't know how to handle a POST request that we
get from GitHub.

This commit introduces an Express.js middleware that mimics the logic
we use in a Ruby server.