Fix current build view (#945)
At the moment we use API V3 to fetch repositories data and along with
repositories we also get data for current_build. Because the payloads
differ between V2 and V3 we need to normalise them in the build
serializer. One thing that was still missing from there was a
pull_request key. We include pull_request key in V2 payloads, but not in
V3 payloads and the build model uses this attribute.