Extract contents of BasicRoute into BaseRouteMixin
This is the first step in a series of refactorings to make the
transition to engines a bit easier.  Because we inherited from
`BasicRoute` in the Application route (and that route referenced another
controller), it was impossible to add a routable engine, as you can't
currently share controllers across the app/engine boundary, only
services/components.

This will also ensure that we're not doing more work than we need to. We
currently opt-in to a lot of behavior and overwrite framework hooks when
it's not entirely clear if that has to happen or not. We'll need to look
into this further to see if it's possible to break down the extracted
`BaseRouteMixin` further into even smaller chunks.

This doesn't actually fix the issue I highlighted with the controller
dependency, it's just the first step in that direction. Because I
converted the previous inheritance implementation, this should behave
identically.