WIP need to figure out how best to include header in engine
Ideally this would be just a part of the application template (which I think should just work, but
it has to wrap the content, not just prepend it).

To that end, I think the solution might be to wrap the header and footer in engines themselves and export the components,
though I am not sure if that only makes them available for the app (and not other engines).

Will have to re-evaluate this later.