Add Dashboard and beta features (#937)
* Current progress of Dashboard feature before merge w/feature-flags work

* Introduce feature flags loaded from server

This is just a current WIP and will need a few modifications before it
can be merged:

Turn off mirage in production.
Do not turn off source maps.
Remove fake entries from mirage/config.js

* Remove duplicate packages from bad merge

* get sorting potentially right

* comment out problematic tests

* Move dashboard feature flag to API response

This fixes some incompatibilities with the two approaches, e.g. weird redirect errors when transitioning to `/dashboard`.

* Update copy: experimental -> beta

* Remove link to Features in dropdown

* Add link to features with terrible sass

Sorry @lislis!

* Add missing passthrough routes to API

* Allow PR env to log to Sentry

* Remove unnecessary logging in Mirage

The issue this helped to debug is already fixed.

* Oops missed one

* Fix broken tests caused by test helper/API incompatibility

The addon we use for our feature flag implementation
(ember-feature-flags) ships with a test helper (withFeature) that we use
to toggle a feature for a given test.

Unfortunately, it isn't compatible with how we were setting feature
flags via the API. We were using the addon's setup method, which would
reset the state of the feature flags, causing our feature flags set in
the tests to be unset and our tests to break.

Given that `setup` is destructive, the fix is to simply enable/disable
based on the API response.  This would still be destructive if there
were overlap between what we return in the mocked out API response and
a feature toggled with the test helper, but we can deal with that later
if/when it arises.

* fix orgs filter styles

* fix dashbaord row styles

* add dashboard feature flag back for dev purposes

* implement (un)star method

* filter repos again

* move actions to controller

* update and refactor star action

* concurrencyfy stars

* change wording when there are no builds

* hide org filter and sync button

* convert dashboard to inline svg

* change some markup

* tweak feature page styles

* add pride feature

* update description, tweak loadin indicator

* add comic sans feature

* Disable dot reporter

This should be temporary! Just to diagnose failures.

* Add logging of PhantomJS version

Another thing to be removed!

* Disable Percy temporarily

Let’s see!

* Disable test temporarily

Also to be removed!

* Revert experimental commits

These changes were an attempt to diagnose/fix mysterious
failures in the Travis environment. We are probably
going to use Chrome instead.

* letter spacing

* Point at staging /beta_features

* Hit correct urls (I think) but getting 404s

* Re-add beta_features endpoint to mirage

I think we still need to add coverage to the feature flags implementation in general, but all acceptance tests will fail if this does not exist.

* Disable dashboard hard-coding

* Try to re-enable sourcemaps

For whatever reason, this has caused build issues in the past, but only on this branch (and not on master) which this now builds off of)

* consider special footer cases

* use simple layout on features page

* add feedback url to feature

* create component for features

* make display of feedback url optional

* change feature names in templates

* do not automatically set dashboard

* add percy snapshot to dashboard

* flexbox dashbaord star and private icon

* also tweak spacing for public star

* tweak star alignment

* Move to blocking loading of feature flags to ensure features loaded before first checks

This is not ideal given the potential for breaking app loading, but should at least guarantee that the app behaves as expected (assuming the endpoint returns data at all).

* Remove errant div

* accomodate loading templates on most pages

* replace simple with center layout

* change hook

* WIP disable failing tests to trigger test deployment

* remove loading indicator from feature toggle

* add error message when feature switch is unsuccessful

* add display name to beta feature model

* sort features by name to avoid jumps in order