Clean auth data from storage before each acceptance test
I run into problems of tests signing in a test user when they were not
supposed to, which usually made the test fail as all the data was not
available (permissions etc).

This commit fixes the problem by always cleaning up the storage data
before the app starts for an acceptance test.