Fix serializers for nested repository records
While working on test's serializers I noticed that when a response
contains nested repository objects (for example when build embeds jobs
with a repository object), the serializer will fail, because it will
mismatch 'repository' type that API sends with our model type in
travis-web - 'repo'. This is not a problem on production at the moment,
because we don't query any endpoints that return such a result, but it's
still good to handle this situation for future proofing serializers.