Try fetching job's config multiple times if it doesn't succeed
Not all of the endpoints return job's config. Because of that we try to
fetch job's config if it doesn't exist. The problem is that till now we
were trying to fetch it only once and we were giving up if the record
was loading. That means that if a config property is accessed while
loading, we will just return it as `undefined` and not attempt a second
try. I haven't noticed this problem in production, but it happened
during tests, so I guess it could happen on production as well.

This commit fixes the situation by retrying until we can do an actual
request.