Fix serializers to work with deep nested relationships
There's a problem in a way we normalize V3 records into JSON API format.
When we extract relationships from a top level resource, we don't
normalize any of the related records, which means that their
relationships are lost. Given a payload like this for a repository:

    {
      @type: 'repository',
      @href: '/repo/1'
      id: 1,
      current_build: {
        @type: 'build',
        @href: '/build/1'
        id: 1,
        jobs: [
          {
            @type: 'job',
            @href: '/job/1
            id: 1
          }
        ]
      }
    }

We would run normalize on the payload, we would loop through the
relationships (`current_build` only in this case) and add relationships
to a `relationships` property. Because we don't run `normalize` on the
`current_build` relationship, we don't get any info about related jobs.

This commit changes the V3 serializer to run `normalize` on related
records, so we essentially end up with a tree of records. After we do
this, we also add all of the records that have a `standard`
representation to the `included` array in order to siedload them into
the store (`relationships` hash is just a set of references to related
records, it isn't supposed to carry extended record info).