Fix broken Github issue link in build title
The Github issue link (see https://travis-ci.org/travis-ci/travis-web/builds/221251257)
was being incorrectly generated because the `owner` attribute
we were using was assumed to be a string, but was actually an object.

This meant we were generating
`https://github.com/[object%20Object]/travis-web/issues/1047` in
production.

I'm not sure what caused this, but applying this fix corrects the links
without breaking other tests.