Add manually-merged changes
This is lacking mirage/serializers/build.js for now,
as that file has changed dramatically and it’s not
immediately apparent to me how to update it. That
means the build stages test is failing, but things
do work as expected, otherwise. I’ll open a new PR
with this and it should work for now.