Fix normalizing pusher payload for repo
We get only slug from pusher, so if pusher payload is received for a
repo that's not already in the store, we will miss `name` and
`owner.login` properties.

This commit fixes the problem by adding the `name` and the `owner`
attributes during normalisation. I also created `RepoV2Serializer` to
not clutter regular V3 serializer with pusher stuff.