Extract builds' template rendering settings to mixin
We currently have the logic required to render builds based on the tab
spread across:

- Repo model (to fetch the actual builds)
- Builds controller (which converts tab state into a call to the Repo
model to grab data)
- AbstractBuilds route (to combine common setup).

The plan is to extract the common behavior of the AbstractBuilds route
into mixins initially, then to replace those settings with more
conventional approaches.

The reason for all of this is my attempt to start build filtering by
type (tied to the open issues surrounding the restructuring of the
`Build History` and `Pull Request` tabs) was stymied by some conventions
that the current approach is laying out.  Specifically, the
`contentType` approach was no longer working with `push` builds.

I'm hoping that by dissecting some of this logic, we can avoid
overbearing/inflexible conventions and make that work more
straightforward.

This commit is just the first step.