Replace Uglify with Babili
My current understanding is that the choice to exclude IE 9 from
the supported browsers has resulted in ES2015 code making it past
the transpliation step, since the supported browsers now all
support some parts of ES2015. Uglify can’t handle ES2015, so
all deployments have failed since that was merged in. This is the
recommended solution I found in this thread:
https://github.com/ember-cli/ember-cli/issues/7006