No longer assume that authentication is always needed
This is a breaking change in terms of implementation, but shouldn't
actually change the behavior of the page (assuming the previous commit
is correct).

The reason this is necessary is that we need to be able to render the
landing page, instead of redirecting before that can happen.

The landing templates should ultimately be moved to the 'index' route
instead of the 'main.index' route to make the setup more conventional,
but this is the state of things for the time being.