[EXPERIMENTAL]: Add ember-route-action-helper
This allows us to skirt controller actions, calling the action of the current route (and bubbling up) instead.
Currently, because our sign out flow is triggered by a component, we set
the `target` property when defining the action.

This would be fine architecturally, but for the fact that we currently
need to affect state within controllers, services and then perform a
redirect.

The alternative that comes to mind is embedding the upcoming router
service, and converting the current controller state to service state, but the routing
service isn't yet public API.

In addition, I think the route-action approach is more intuitive/common
than setting the action target.

The current plan would then be to trigger an action on the route
(bubbling up to be handled within the ApplicationRoute), which would
internally call the auth service where possible. The route can still
make `controllerFor` calls in the interim until the transition from
controllers to services is complete. Afterwards, it might be better to
return to centralizing this within the auth service and utilizing the
target, but that's debatable.

I've made a note to discuss this in an upcoming :ketchup:.