Add default for pro-version feature flag (#1140)
The lack of default for `pro-version` was causing the setting
of the flag for a test to bleed through to future tests, since
it wasn’t being reset when the application was created. This
was producing false negatives in our snapshot comparisons,
particularly around the presence of the open source `cta`
header. It’s my hope that these will stop happening now.