Randomize loading logo (#1180)
* Show random logo on app boot

This shows a random logo (from all the SVG variants I noticed)
when the user first boots up the app.

Implementation-wise, it might be confusing at first as to why the
random-logo service was added instead of making this a part of the
`loading-screen` component. The reason is that we utilize several layers of loading
templates which -- were the randomization done at the component level --
would mean that the logo might jump between available options. This
would be jarring, so I opted for making the selection when the service
is first initialized and then referencing that state whenever the
component is used.

* Exclude logos from fingerprinting