Revert chunked subscription to Pusher channels (#1185)
My unfamilarity with this code caused problems in the pro environment,
where subsequent chunk subscriptions failed because of incorrect
authorisation information. I’m going to fully revert for now.

The general idea of the problem is that the `bulk_ajax` method fetches
authorisation information for each channel in bulk, but since it happens
synchronously, it was only fetching it for the first chunk of channels.
I’d like to change it to fetch for each chunk.