Fix tests
The problem was a logical disconnect between the requirements of two tests.
The sidebar does not show inactive repositories, but we simultaneously need to redirect to the `getting_started` page if there are not any repositories.
The setup of one test visited the repo.index route with a single,
inactive repository created. The sidebar was redirecting (correctly) to
the getting started page as a result.

The change was to add some brittleness to the sidebar's logic to handle
this, basically exempting the redirect unless we're on the index page.