Don't always fetch stages info (#1196)
We need stages info only on a build page and then we can fetch it
lazily, so we don't need to always include it.

This commit changes the stages relationship into an async relationship,
so now we don't have to always attach stages info, but fetch it only
when needed. Fetching is done using  the `/build/:build_id/stages`
endpoint with help of the `groupRecordsForFindMany` function in
the stages adapter - its usage allows us to group multiple stages
requests into one request that's done with `findMany`.