Implement more flexible filtering for the store
Ember Data's filtering has rather limited capabilities. The biggest
problem is that it will check if a record should be added to a filterred
array only when a record is added to the store or a property of a record
changes. This means that it will not refresh any records in a filtered
array if you want to depend on anything other than properties defined on
the record, which may be limiting in some scenarios.