Change test randomisation to not run with Percy (#1238)
I did manage to fix a lot of the spurious visual diffs by [eliminating some test bleed-through](https://github.com/travis-ci/travis-web/pull/1140/files) (and we got some [welcome help for others](https://github.com/travis-ci/travis-web/pull/1147)), but we still have one or two on each build since then. I don’t have time at the moment to look deeply enough to remove the rest of the bleed-though, so instead I’m suggesting that we stop randomising tests when we’ll be producing diffs with Percy.

I don’t love it, but the noise from the false diffs is grating and I don’t know if/when I’ll be able to address the cause.