Remove status link in test environment (#1240)
My reason for this is to reduce false negatives in Percy. The assertions
didn’t add much value because the appearance of the status section in
the footer is dependent on the presence of the status API URL, which
isn’t something that was tested in those acceptance tests. There’s a
component test that mocks the API, which seems sufficient.