Improve switch accessibility (#1229)
The main change here is to make the switches into buttons
rather than anchor tags, which automatically makes them
behave more expected and is more semantically accurate.

There’s an unfortunate side effect here that the rendering
of the switches includes a small amount of extra whitespace
on the side in Firefox 45 ESR. This causes the Percy renderings
to be slightly off, but the browser isn’t widely used according
to recent statistics. I couldn’t find any CSS to make the whitespace
disappear, so I’m grudgingly merging this.