Update pagination data on each reload
When forceReload option is set to true, we will run a query each time
the `store.paginated` function is executed. We may still return
immediately, but the query will run in the background anyway. This
commit updates LivePaginatedCollection to always update pagination data
on subsequent runs when forceReload is set to true.