Make getting back to dashboard instantaneous
By using an updated version of store.filter with forceReload turned on
for accounts, we can avoid waiting for any load when a user navigates
away from the dashboard and then returns. A query to refresh data will
be still send in the background, but the promise will resolve
immediately.

This commit also removes a limit on repos so we get the default from the
API.