Ensure we yield the child task
This ensures that the parent task is considered as `isRunning` until the child task completes.
This was the last piece, as previously the modal was still visible with the Trigger Build button no longer showing a loading status.
In theory, this meant that it could be clicked a second time.