Refactor trigger build logic to use derived states (#1256)
* Refactor trigger build logic to use derived states

This extracts out the tasks into separate parent/child tasks and uses
`timeout` instead of Ember.run.later.

* Fix bug caused by using `firstObject`

`firstObject` is returning a cp instead of -- you might be guessing it -- the first object. So we will just use JS instead.

* Ensure we yield the child task

This ensures that the parent task is considered as `isRunning` until the child task completes.
This was the last piece, as previously the modal was still visible with the Trigger Build button no longer showing a loading status. In theory, this meant that it could be clicked a second time.