Fix tooltip width and hover behavior on dashboard
Instead of messing with the styles directly, I just tried to
modify the DOM here to match what we've done elsewhere. I also changed
the nesting of the link (to exclude the span icons), which seems to have
stopped the previous underline behavior, where we were underlining the
icons as well as the link itself.