Update GitHub icons (#1199)
* Support live reloading styles by configuring sass extension

* Add external-link-to component

This will ensure that we follow security best-practices without having
to think about it. Tests will also fail if we try to make it insecure.

More details about using "noopener noreferrer" can be found here:
https://www.jitbit.com/alexblog/256-targetblank---the-most-underestimated-vulnerability-ever/

* Update icons

* Allow class to be passed through to link

* Convert build-header elements to use new styles/component

* Modify branches icons (with hacked together sass)

* Modify build history build item display

* Fix styling of running items (yellow instead of gray)

* Use fill helper instead of opacity because I am terrible at CSS

* Fix build status display in build-header

* Fix broken tests

* Fix tests now that # is in an icon

* Update titles

* Fix broken test after rebase

* Center tooltip text

The tooltip seems to be as large as the span itself (not the text containing it), which makes centering the text necessary

* Remove duplicate `title` attribute for `data-tooltip` instead

* Fix broken tests asserting that `title` is set

* Restore the previous title text (my mistake)

* Revert assertions to make tests pass

* Update tooltips for build header (#1254)

* special case build ehader tooltips

* change tooltip text to View__

* change to new icons on dashboard

* Final fix-ups

* All the icons (#1268)

* fix gh icons on owner page

* fix icons on build history/ pull reuqest/ branches page

* fix underline on build history/ pull requests

* Fix remaining tests

* Fix GitHub link to sha on dashboard

* Fix tooltip width and hover behavior on dashboard

Instead of messing with the styles directly, I just tried to
modify the DOM here to match what we've done elsewhere. I also changed
the nesting of the link (to exclude the span icons), which seems to have
stopped the previous underline behavior, where we were underlining the
icons as well as the link itself.