Remove custom pusher authentication code
We were using this code to allow authenticating multiple channels at
once, because a user had to subscribe to as many channels as
repositories they own. Which could mean thousands for some users. Now we
only subscribe a user to one channel and then sometimes to repo channels
as needed, so this logic is no longer necessary.