Save subscribed channels in TravisPusher
In tests we sometimes need to check if a subscription to a channel has
been made. This is impossible, because we don't really connect to Pusher
in testing mode.

This commit saves subscribed channels into active_channels array for
later use in tests.

A better way to do it would be to provide a separate implementation for
tests (possibly by using some kind of adapters pattern), but it might be
hard to do given current setup, so I went for the simpler thing for now.