Subscribe to public repos on pro too
When a user opens a repo that they aren't a collaborator of, we will
subscribe them to a public repo channel, so they still get updates for
the repo. This was done only on .org, but now that we plan to merge the
platforms it should work on .com version too.

The code needed to be changed a little bit, because now we need to also
care about private repositories and also we don't want to subscribe to
repositories that are already covered on user channel (ie. if the
current user is a collaborator, they would get an update through a user
channel anyway).