Improve stage allowed failures text (#1194)
This updates the allowed failures text so it doesn’t incorrectly state that the next stage will continue even when one of the not-allowed-to-fail jobs failed. Thanks to @webknjaz for the [bug report](https://github.com/travis-ci/beta-features/issues/11#issuecomment-306060542).

It also adds text about allowed failures on the last stage, without mention of continuing to another stage. Thanks to @timofurrer for the [bug report](https://github.com/travis-ci/beta-features/issues/11#issuecomment-308727991).