Fix subscribing to private jobs (#1302)
Subscription to private jobs was recently changed in a way that we allow
to subscribe to both private and public jobs, regardless of the
platform. This check requires access to a `repo.private` property and
unfortunately in certain cituations the `job.subscribe()` method may be
called when job is still not fully loaded (meaning that repository won't
be loaded either).

This commit fixes the situation by waiting to load the job before
subscribing. This is a bit hacky and we should probably fix the app to
not even reach this place when the job is not loaded, but this was the
best way to do it at the moment.

I also removed changed bulk_ajax authorizer to travis_ajax authorizer.
We don't need to authorize multiple channels anymore, so I could
simplify the authroizer and thanks to that also remove some code from
subscribing logic (no need for explicit ajax call when subscribing for a
log channels anymore).