Split new SSH key test from deletion test
This was failing with “attempted to handle event didSetProperty”,
presumably because of the way that the SSH key always uses the
ID of its repository. This error doesn’t happen in production,
though. Splitting the tests and faking the non-existence of
the key in the API is sufficient to test what’s intended.