Convert id in stages adapter findRecord to int
Ids returned from the API are ints, so we need to convert the id to int
in order to properly fetch the stage by id. The tests didn't catch this,
because Mirage was returning ids as strings.

This commit fixes the bug by converting the id to int before filtering,
but also changes tests to also return id as int to prevent these kind of
problems in the future