Change headers to spans on dashboard row
In theory each section should include a header, but I don't think that
this is a good practice in case of repository-row. When you open a list
of headings of the dashboard page, where the number of repositories is
bigger than 100 (100 active repos + starred repos), the list of headers
is huge (at least a few hundred), which makes it unusable in my opinion.