Only apply settings switch state locally if successfully persisted (#1336)
* Only toggle the property if the request was successful

The user will be shown an error flash message if something goes wrong.

* Remove loading indicator

This is out of place in the current UI and is shown to the right of the text

* Fix failing tests

Previous implementation relied on changing local state first and then attempting to persist that state. Instead, try to persist the new state and only apply locally if successful.