Don't update records with stale data
Data updates doesn't arrive to the browser in any specific order.
Because of that it's possible that data in an update delivered by
Pusher is older than data that we have available locally.

This commit overrides `store._pushInternalModel` function in order to
guard against such situations by checking the `updatedAt` attribute. If
a record that is being pushed to the store is older than a record that
we already have, we will just return the existing record.