Fix sign out redirect (#1404)
When a Travis token expires and a user tries to open the app, we will
sign them out. The problem is that due to changes in e71862c we're no
longer running some of the sign out steps when the sign out is initiated
from the `auth` service. This commit gets back to calling the
`afterSignOut` every time a `signOut` is initiated.