Highlight the active organization on the sidebar (#1480)
With this change now the user and organization avatars are included inside the "link-to" ember component, which is essentially an anchor tag used as a container for the whole list item. Instead of images being outside of the anchor and sharing a wrapper div with the link component.

This way, the whole list item is clickable and can be set as active/inactive, html structure is more consistent as we are removing the unused wrapper divs, also avatars now have borders when they are selected.