WIP: Use  url param instead of offset
Users were able to manually specify the offset in the URL, which led to rounding errors in the rendered pagination links.

This makes it simpler to use from a user's perspective and also avoids exposing something the user should not be able to influence (yet). We can always re-evaluate this later if users would find it useful, as API still supports.