Fix status icon flickering in loading template footer(s)
Depending on the loading behavior, users can see a flicker in the footer
while the {{travis-status}} component is re-rendered. Currently, in production, we make
3 requests to the status endpoint while trying to render the "Build
History" tab from scratch. The reason is our use of loading templates
renders the components three separate times (for a unifying feel
across route transitions). The downside is that this behavior has side
effects for both aesthetics and, potentially, performance.

Instead, this moves the request to an ember-concurrency task on a new
`app-loading` service which survives across component invocations. The
task automatically drops subsequent task invocations while the component
itself simply renders the first successful call.