Add subscription checkmarks for all accounts
I should never have had it fetching /subscriptions whenever
the account was changing, it should have been up in the accounts
route as it is now.

It’s kind of weird to use the subscriptionError client-side
attribute this way, but it seems better than making yet another
weird compound model that has to be specially handled elsewhere.