Remove LimitedArray for forcing singleton flash (#1653)
In looking at updating to Ember 3, I ran into some trouble because the way the
flashes service works involves mutating an arranged ArrayProxy, which is now
forbidden. LimitedArray must have been used elsewhere in the past, maybe with
the pseudo-infinite scroll that used to exist? It seemed fine to me to convert
to only display the first flash message instead.