Update to use V3 for requests (#1640)
V3 has different semantics for the request-build relationship:
it went from one-to-one to a request having many builds. The
reality remains that a request only has one build, though. If
further handling of this relationship is needed, there should
probably be an alias to the first build on the request model.