Remove content security policy
I’ve found that CSP warnings cause a lot of confusion and noise,
obscuring the true problems we’re trying to track down. The
“report only” nature of the log messages doesn’t stand out to
people, I think they see the red text and assume it’s a problem.
Until we get to a point where we’ve actually safelisted everything
and stop having false negatives, I’m removing this.