Update Ember CLI to 3.4 (#1784)
This doesn’t include template linting, which has been added to the
blueprint, but is Too Much™ to cover at the moment.

It’s unfortunate that the timeago jQuery plugin has to be inlined
like this but I couldn’t find a better way now that jQuery is no longer
global. Hopefully we can just extract the relevant code when we’re
ready to abandon jQuery.