Extend test coverage to cover recent regression
A user reported in https://github.com/travis-ci/travis-ci/issues/10112
that we were overwriting spaces when colored text was printed to the
console in logs.

I wasn't able to reproduce, but wanted to extend our current coverage in
to keep us from accidentally breaking this in the future.

As part of this work, I have added a new page object helper to easily
reference the entire text of a given line. Including this unfortunately
required modifying the scope of the page object's collection, so this
ended up being a bit more invasive than I would have liked, but it seems
worth it given the ability to assert on more complex lines in the future.