Hide Trigger Build link for migrated repositories on .org
Note: in regards to the removed test, we used to test unauthenticated users as well,
which is totally reasonable. However, while working
through this, it seems that we aren't actually unauthenticated as far as
mirage is concerned, as an auth token
is still being passed (otherwise we would get a 403 based on the current
config). So this test became an exact copy,
where we would simulate being logged out by not passing the correct
permissions. The only reason the other test passed
was because we were randomly not using an `andThen` so when we checked
that the element was not present, it wasn't.

We weren't, however, testing the actual logic of this page.