Add YAML view (#1826)
This replaces the JSON shown on the config tab with the original .travis.yml
file. It adds a config tab for builds; one only showed on jobs previously.
Config messages are moved to the config tab, but they still are only
produced for selected organisations.

.travis.yml files are not stored before June 2018; in that situation, the
API returns a fake YML error file, which manifests in the UI as a “disabled”
config tab with a tooltip.