UI updates to show migration status and build history (#1889)
* Convert repo model to use new `migrationStatus` attr from API

* Add method to generate link to travis-ci.com

* Render link to travis-ci.com for already migrated repository

* Only show migrated button on .org

* Add test ensuring link to .com is shown

* Show banner on .com with link to build history for migrated repository

* Fix linting errors

* Disable settings page for migrated repositories on .org

* Hide Trigger Build link for migrated repositories on .org

Note: in regards to the removed test, we used to test unauthenticated users as well,
which is totally reasonable. However, while working
through this, it seems that we aren't actually unauthenticated as far as
mirage is concerned, as an auth token
is still being passed (otherwise we would get a 403 based on the current
config). So this test became an exact copy,
where we would simulate being logged out by not passing the correct
permissions. The only reason the other test passed
was because we were randomly not using an `andThen` so when we checked
that the element was not present, it wasn't.

We weren't, however, testing the actual logic of this page.

* Add notice to current tab of .org when viewing migrated repository

* Fix linting error

* Remove spaces :scream:

* Show banner on .org linking to migrated repository on .com

* Ensure banner is below the repo header

* Restore behavior from mistakenly overwritten PR

This restores the behavior from https://github.com/travis-ci/travis-web/pull/1894