Remove ember-decorators (#1910)
We started using this for the auto-getting of dependent keys,
which it no longer supports. Rather than be stuck on an older
version, we decided to move away from it altogether.

Most of the transformations were done with this:
https://github.com/travis-ci/ember-abandon-decorators-codemod

I did some hand-tweaking when necessary to preserve whitespace
and comments and shorten some variable-gettings.