Update generation of redirect URL (#1950)
Redirection is broken from non-root paths, like /dashboard. The
generated redirect URL is like travis-ci.orgundefined.

It’s possible this broke with the Ember 3.7 update. For expediency,
I’ll merge this with no accompanying test but spend some time
trying to understand why it was written as it was and adding
an automated test.