Revert to closer-to-previous behaviour
I’m not sure what the reason is for constructing the return
URL in this way, but it worries me to just eliminate it.

The transition.intent.url interface isn’t public but I
don’t know how else to get at this in a way that will work
properly in the test environment.