Convert more tests to modern testing style (#2054)
* Convert broadcasts test to new style

* Convert home/with-repositories to new testing style

* Fix ESLint error

* Convert  another test

* Silence ember-keyboard-shortcuts deprecation

* Convert owner not found test

* Update another test

* Convert enterprise banner test

* Convert enterprise/navigation test

* Ensure user is signed out in navigation test

This causes test flappiness when running all tests together

* Convert repo active on org test

* Convert automatic sign out test

This also includes a new helper (credit to @ppcano here: https://github.com/emberjs/ember-test-helpers/issues/332#issuecomment-414641697) to work around https://github.com/emberjs/ember-test-helpers/issues/332.

* Fix ESLint errors

* Convert sign out test

* Convert feature flags test

* Convert pull requests test

* Convert builds restart test

* Convert builds view pull request

* Convert invalid builds test

* Convert build stages test

* Convert search flow test

* Convert profile not found test

* Convert profile redirect test

* Convert profile billing test (phew)

* Convert no search results test

* Convert user settings test

* Convert basic layout test

* Convert profile filtering test

* Conver view token test

* Convert test and remove unsightly unless/else

* Convert default layout test

* Convert header layout test

* Convert cta layout test

* Convert layout/plans test

* Convert job build matrix test

* Convert jobs/restart test

* Convert repo trigger build

* Convert invalid job log test

* Convert job test and fix resulting errors

* Convert job log error test

* Convert job delete test

* Convert job debug test

* Convert job cancel test

* Convert repo/not found test

* Convert repo/requests test

* Whoops missed the helper file

* Convert build list routes test

* Convert repo branches test

* Convert repo caches test

* Fix helper import in page object

* Convert repo show test

* Convert public pusher test

* More conversions

* Clear local storage between tests as well

* Missed part of the test

* Convert first sync test

* Add ember-lifeline

Hoping based on https://engineering.linkedin.com/blog/2018/01/ember-timer-leaks that it will help us identify some leakage that is causing some annoying test flappiness.

* Add data-test attrs to be able to use waitFor helper

* Convert sync test and remove unnecessary timeout/corresponding config

* Revert "Add ember-lifeline"

Was just an experiment, removing for now.

This reverts commit 9e434c4330fb79fae594ca8f86bc014add049597.

* Fix ESLint errors

* Convert skipped test (it is still skipped)

* Remove use of custom waitForElement helper

This is now built-in as `waitFor` within @ember/test-helpers.

* Convert environment instance initializer unit test

* Convert pro environment test and fix formatting in another test

* Import correct initialize function in test

* Skip sync tests temporarily

It appears that these tests are very flaky, so need to figure out a way to permanently fix them. I do not want them, however, to hold up this chunk of work.

* Remove legacy test helpers

* Add space for formatting

* Convert "unit" test to integration test

This does lose a single assertion that we are calling the auth service correctly, but I would argue that this should be a unit test on the auth service if it is needed at all.

* Fix ESLint errors

* Add comment regarding slow acceptance test

* Move final tests within module block

This was causing a super weird error regarding `window.wait`.

* Remove unnecessary import

We no longer use `withFeature` but rather import `enableFeature` directly in any test that needs it

* Remove logging statement from template

* Use our own setupApplicationTest helper

This will automatically sign a user out and clear browser state, ensuring every test starts off with the same state

* Remove unnecessary storage calls

This already happens within `setupApplicationTest`.

* Remove unnecessary use of signOutUser helper

I have left it in place for now (in case it has some use later), but we reset the local/session storage anyway before each acceptance test, and our only uses of it right now were essentially duplicating that work.